+++ Source/Class/TypeScriptESBuild.ts
@@ -7,9 +7,9 @@
 export default new (await import("commander")).Command()
 	.name("TypeScriptESBuild")
 	.version(process.env["VERSION_PACKAGE"] ?? "0.0.1")
-	.description("Builds files")
-	.argument("<Files...>", "Files to build")
-	.option("-es, --ESBuild <File>", "esbuild configuration file")
-	.option("-ts, --TypeScript <File>", "TypeScript configuration file")
+	.description("🌀 Build TypeScript")
+	.argument("<File...>", "Build File.")
+	.option("-es, --ESBuild <File>", "ESBuild Configuration.")
+	.option("-ts, --TypeScript <File>", "TypeScript Configuration.")
 	.action((await import("../Function/Build.js")).default)
 	.parse();