+++ Source/Function/Build.ts
@@ -51,39 +51,6 @@ export default (async (...[File, Option]: Parameters<Type>) => {
 	);
 
 	Exec(`tsc -p ${Option?.TypeScript ?? "tsconfig.json"}`);
-
-	Exec(
-		[
-			"typedoc",
-			"--commentStyle all",
-			"--gitRevision main",
-			`--customCss ${resolve(`${Current}/../Stylesheet/Theme.css`)}`,
-			"--includeVersion",
-			"--out ./Documentation",
-			`--plugin ${resolve(`${Current}/../../Target/Variable/Load.js`)}`,
-			"--plugin typedoc-plugin-remove-references",
-			"--plugin typedoc-plugin-rename-defaults",
-			"--plugin typedoc-plugin-mdn-links",
-			"--plugin typedoc-plugin-zod",
-			"--plugin typedoc-plugin-merge-modules",
-			"--plugin typedoc-plugin-keywords",
-			"--searchInComments",
-			`--keywords ${
-				(
-					await (
-						await import("../Function/JSON.js")
-					).default("package.json", process.cwd())
-				)?.keywords?.join(" --keywords ") ?? " typescript-esbuild "
-			}`,
-			"--theme TypeScriptESBuild",
-			"--entryPointStrategy expand",
-			"--mergeModulesRenameDefaults",
-			"--mergeModulesMergeMode module",
-			`--entryPoints ${Object.values(Configuration.entryPoints).join(
-				" --entryPoints ",
-			)}`,
-		].join(" "),
-	);
 }) satisfies Type as Type;
 
 import type Type from "../Interface/Build.js";